reorganize pages
