+++ src/App.jsx
@@ -4,8 +4,6 @@ import ScrollToTop from './ScrollToTop';
 import Navbar from './components/General/Navbar';
 import Footer from './components/General/Footer';
 import Home from './pages/Home';
-import About from './pages/About';
-import Events from './pages/Events';
 import Blog from './pages/Blog';
 import Archive from './pages/Archive';
 import NotFound from './pages/NotFound';
@@ -20,8 +18,6 @@ function App() {
 			<ScrollToTop />
 			<Routes>
 				<Route path='' element={<Home />} />
-				<Route path='about' element={<About />} />
-				<Route path='events' element={<Events />} />
 				<Route path='blog/:blogId?' element={<Blog />} />
 				<Route path='*' element={<NotFound />} />
 				<Route path='archive' element={<Archive />} />
