+++ src/pages/Archive.jsx
@@ -3,6 +3,9 @@ import EventInfoItem from '../components/Archive/EventInfoItem';
 import { archiveData } from '../data/archive';
 import '../styles/Archive.css';
 import useTitle from '../components/General/useTitle';
+import PhotoCarousel from '../components/Archive/PhotoCarousel';
+import 'slick-carousel/slick/slick.css';
+import 'slick-carousel/slick/slick-theme.css';
 
 export default function Archive() {
 	useTitle(' | Archive');
@@ -18,6 +21,7 @@ export default function Archive() {
 			<div className='archive-header'>
 				<h1>Workshop Archive</h1>
 			</div>
+			<PhotoCarousel />
 			{data.map((quarter, index) => (
 				<div key={index} className='quarter'>
 					<h2>{quarter.quarter}</h2>
