+++ src/styles/Home.css
@@ -24,12 +24,16 @@
 /* ========== Banner.jsx ========== */
 .banner-container {
 	/* background-color: var(--bgPink); */
-	background: linear-gradient(180deg,rgba(255, 178, 221, 1) 60%, rgba(255, 255, 255, 1) 100%);
+	background: linear-gradient(
+		180deg,
+		rgba(255, 178, 221, 1) 60%,
+		rgba(255, 255, 255, 1) 100%
+	);
 	display: flex;
 	justify-content: center;
 	padding: 0;
 	width: 100%;
-	margin-top: calc(-1 * var(--navbar-height)); 
+	margin-top: calc(-1 * var(--navbar-height));
 	padding-top: var(--navbar-height);
 }
 
@@ -46,15 +50,15 @@
 
 #light-group.light-glow {
 	fill: url(#glow-gradient);
-  filter: url(#glow-effect);
+	filter: url(#glow-effect);
 }
 
 #wire-group.wire-glow {
-  filter: url(#glow-effect);
+	filter: url(#glow-effect);
 }
 
 .off {
-  fill-opacity: 0.5;
+	fill-opacity: 0.5;
 }
 
 .on {
@@ -63,20 +67,20 @@
 }
 
 .blink {
-  animation: blinkAnimation 1s steps(1, start) infinite;
+	animation: blinkAnimation 1s steps(1, start) infinite;
 }
 
 @keyframes blinkAnimation {
-  0% {
-		fill: #A164BE;
-  }
+	0% {
+		fill: #a164be;
+	}
 	50% {
 		fill: white;
 		filter: url(#terminal-glow);
 	}
-  100% {
-		fill: #A164BE;
-  }
+	100% {
+		fill: #a164be;
+	}
 }
 
 /* ========== HackDescription.jsx ========== */
